fix(TST-09): align shanty animation timers with time

+++ src/lincity/modules/shanty.hpp
@@ -128,10 +128,10 @@ public:
     virtual void bulldoze() override;
     void makeFire();
 
-    int anim;
+    unsigned long anim;
     bool start_burning_waste;
     std::list<ExtraFrame>::iterator waste_fire_frit;
-    int waste_fire_anim;
+    unsigned long waste_fire_anim;
 };
 
 #endif // __LINCITYNG_LINCITY_MODULES_SHANTY_HPP__
